# Conflicts:
#	build.gradle
#	src/main/kotlin/me/steven/indrev/networks/fluid/FluidNetwork.kt
#	src/main/kotlin/me/steven/indrev/networks/item/ItemNetwork.kt

+++ src/main/kotlin/me/steven/indrev/networks/Node.kt
@@ -3,7 +3,7 @@ package me.steven.indrev.networks
 import net.minecraft.util.math.BlockPos
 import net.minecraft.util.math.Direction
 
-data class Node(val origin: BlockPos, val target: BlockPos, val dist: Short, val direction: Direction) : Comparable<Node> {
+data class Node(val origin: BlockPos, val target: BlockPos, val dist: Int, val direction: Direction) : Comparable<Node> {
     override fun compareTo(other: Node): Int {
         return when {
             this.dist > other.dist -> 1
